Reorg more pages out of "more" subpath
My rationale for this is that URLs are generated based on the directory
structure of the Markdown files. I'm not sure if there's a way around
this with Quarto. I don't want URLs like
`example.com/more/garden/2024.html` or
`example.com/more/cognitive_load.html`. I simply wanted these items to
be under the "More" dropdown in the navbar! After moving the pages, the
URLs will be (IMO) more sensible.